% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqFit.R
\name{chisqFit}
\alias{chisqFit}
\title{Calculate model fit}
\usage{
chisqFit(resObserved, data, ncr = NULL, DstarM = FALSE, tt = NULL,
  m = NULL, by = NULL)
}
\arguments{
\item{resObserved}{output from \code{\link{estObserved}}.}

\item{data}{Either a dataframe containing data used for estimating the models or a matrix containing custom densities used for estimating.}

\item{ncr}{The number of observations per condition-response pair. This is used to weight the objective function. Should only be supplied when data is a matrix containing custom densities.}

\item{DstarM}{Logical. Should the DstarM fit measure be calculated or the traditional fit measure?}
}
\description{
Calculate model fit
}
\details{
This function allows a user to manually calculate a chi-square goodness of fit measure for a model.
This is useful for comparing a traditional analysis and a D*M analysis. For completion, this function can also calculate a
D*M fit measure. We do not recommend usage of the D*M measure. While the chi-square fit measure is
identical to the value of the optimizer when fitting, the DstarM fit measure is not equal to that of a DstarM analysis.
This is because this function calculates the DstarM fit measure on the complete distribution, not on the
model distributions, as is done during the optimization.
}
\examples{
tt = seq(0, 5, .1)
pars = c(.8, 2, .5, .5, .5, # condition 1
        .8, 3, .5, .5, .5,  # condition 2
        .8, 4, .5, .5, .5)  # condition 3
pdfND = dbeta(tt, 10, 30)
# simulate data
data = simData(n = 3e3, pars = pars, tt = tt, pdfND = pdfND)
# define restriction matrix
restr = matrix(1:5, 5, 3)
restr[2, 2:3] = 6:7 # allow drift rates to differ
# fix variance parameters
fixed = matrix(c('sz1', .5, 'sv1', .5), 2, 2)
\dontrun{
# Run D*M analysis
res = estDstarM(data = data, tt = tt, restr = restr, fixed = fixed)
coef(res)
summary(res)
}

}

