% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSter.R
\name{getSter}
\alias{getSter}
\title{Estimate variance of nondecision density}
\usage{
getSter(res)
}
\arguments{
\item{res}{An object of class \code{D*M}.}
}
\description{
Estimate variance of nondecision density
}
\details{
The object \code{res} can either be output from \code{estDstarM} or output from \code{estND}.
if the former is supplied, \code{getSter} attempts to calculate the variance of the
nondecision distribution by subtracting the variance of the model distribution from the
variance of the data distribution. If the latter is supplied, the variance is calculated by
integrating the nondecision distribution.
}
