% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTer.R
\name{getTer}
\alias{getTer}
\title{Calculate Mean of the nondecision distribution.}
\usage{
getTer(res, dat)
}
\arguments{
\item{res}{An object of class D*M.}

\item{dat}{The data object used to create \code{res}.}
}
\value{
A vector containing estimates for the mean of the nondecision densities.
}
\description{
Calculate Mean of the nondecision distribution.
}
\details{
The object \code{res} can either be output from \code{estDstarM} or output from \code{estND}.
If the former is supplied it is also necessary to supply the data used for the estimation.
The mean will then be estimated by substracting the mean of the model densities from the mean of the data density.
If the latter is supplied than this is not required; the mean will be calculated by
integrating the nondecision distribution.
}
