% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsQuantiles.R
\name{obsQuantiles}
\alias{obsQuantiles}
\title{Calculate model fit}
\usage{
obsQuantiles(data, probs, what = "cr")
}
\arguments{
\item{data}{A dataframe with: a column named \code{rt} containing response times in ms,
a column named \code{response} containing at most 2 response options, and an
optional column named \code{condition} containing a numeric index as to which conditions
observations belong.}
}
\description{
Calculate model fit
}
\examples{
tt = seq(0, 5, .01)
pars = c(.8, 2, .5, .5, .5, # condition 1
        .8, 3, .5, .5, .5,  # condition 2
        .8, 4, .5, .5, .5)  # condition 3
pdfND = dbeta(tt, 10, 30)
# simulate data
data = simData(n = 3e3, pars = pars, tt = tt, pdfND = pdfND)
probs = seq(0, 1, .01)
obsQuantiles(data)
matplot(tt, obsQuantiles, type = 'l')
}

