% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDescriptives.R
\name{rtDescriptives}
\alias{rtDescriptives}
\title{Descriptives of reaction time data}
\usage{
rtDescriptives(data, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{A reaction time dataset. Must be a dataframe with $rt, $condition
and $response.}

\item{plot}{Logical, should a density plot of all condition-response pairs be made?}

\item{verbose}{Logical, should a table of counts and proportions be printed?}
}
\value{
Invisibly returns an object of class 'D*M'. It's first element is \code{table} and contains raw counts and proportions for
condition response pairs, conditions, and responses. It's second element \code{plot} contains a ggplot object.
}
\description{
Descriptives of reaction time data
}
\details{
This function and \code{\link{rtHist}} are helper functions to inspect raw data.
}
\examples{
tt = seq(0, 5, .01)
pars = matrix(.5, 5, 2)
pars[1, ] <- 1
pars[2, ] <- c(0, 2)
dat = simData(n = 3e3, pars = pars, tt = tt, pdfND = dbeta(tt, 10, 30))
x = rtDescriptives(dat)

print(x$table, what = 'cr')
print(x$table, what = 'c')
print(x$table, what = 'r')
}
