% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{upgradeDstarM}
\alias{upgradeDstarM}
\title{Upgrade a DstarM object for backwards compatibility}
\usage{
upgradeDstarM(x)
}
\arguments{
\item{x}{an object of class \code{D*M} or \code{DstarM}.}
}
\value{
An object of class \code{DstarM.fitD}, \code{DstarM.fitND}, or \code{DstarM.fitObs}.
}
\description{
Upgrade a DstarM object for backwards compatibility
}
