% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize two pdfs}
\usage{
normalize(x, tt, props = NULL)
}
\arguments{
\item{x}{a numeric matrix or vector where each column represents a probability density function evaluated at the grid defined in \code{tt}.}

\item{tt}{a numeric grid defined in \code{x}.}

\item{props}{the value each density should integrate to.}
}
\description{
Normalize two pdfs
}
